+++ extensions/purelymail/src/list-domains.tsx
@@ -9,110 +9,67 @@ import {
   Alert,
   confirmAlert,
   showToast,
-  Toast,
-  popToRoot,
+  Toast
 } from "@raycast/api";
-import { getFavicon } from "@raycast/utils";
+import { getFavicon, useCachedState } from "@raycast/utils";
 import { useEffect, useState } from "react";
 
 import { deleteDomain, getDomains } from "./utils/api";
 import { Domain, Response } from "./utils/types";
 import ErrorComponent from "./components/ErrorComponent";
 
-interface State {
-  domains?: Domain[];
-  showDetails: boolean;
-  error?: string;
-  isLoading?: boolean;
-  listSharedDomains?: boolean;
-}
-
 export default function ListDomains() {
-  const [state, setState] = useState<State>({
-    domains: undefined,
-    showDetails: false,
-    error: "",
-    isLoading: false,
-    listSharedDomains: false,
-  });
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState("");
+  const [domains, setDomains] = useCachedState<Domain[]>("domains");
+  const [isShowingDetails, setIsShowingDetails] = useState(false);
+  const [filter, setFilter] = useState("");
 
-  useEffect(() => {
-    async function getFromApi() {
-      const response: Response = await getDomains(true);
+  async function getFromApi() {
+    setIsLoading(true);
+    const response: Response = await getDomains(true);
 
-      switch (response.type) {
-        case "error":
-          setState((prevState) => {
-            return { ...prevState, error: response.message, isLoading: false };
-          });
-          break;
-
-        case "success":
-          setState((prevState) => {
-            return { ...prevState, error: "", domains: response.result.domains };
-          });
-          break;
-
-        default:
-          setState((prevState) => {
-            return { ...prevState, isLoading: false };
-          });
-          break;
-      }
+    if (response.type==="error") {
+      setError(response.message);
+    } else {
+      setDomains(response.result.domains);
+      await showToast({
+        title: "SUCCESS",
+        message: `Fetched ${response.result.domains?.length} domains`
+      })
     }
+    setIsLoading(false);
+  }
+  useEffect(() => {
     getFromApi();
   }, []);
 
   const toggleShowDetails = () => {
-    setState((prevState) => {
-      return { ...prevState, showDetails: !state.showDetails };
-    });
+    setIsShowingDetails(!isShowingDetails);
   };
 
   const handleDelete = async (domain: string) => {
     if (
       await confirmAlert({
         title: `Delete ${domain}?`,
+        icon: { source: Icon.DeleteDocument, tintColor: Color.Red },
         message: "You will not be able to recover it",
         primaryAction: { title: "Delete", style: Alert.ActionStyle.Destructive },
       })
     ) {
-      setState((prevState: State) => {
-        return { ...prevState, isLoading: true };
-      });
+      setIsLoading(true);
 
       const response = await deleteDomain(domain);
-      switch (response.type) {
-        case "error":
-          await showToast(Toast.Style.Failure, "Purelymail Error", response.message);
-          setState((prevState) => {
-            return { ...prevState, isLoading: false };
-          });
-          break;
-
-        case "success":
-          setState((prevState) => {
-            return { ...prevState, isLoading: false };
-          });
-          await showToast(Toast.Style.Success, "Domain Deleted", "DOMAIN: " + domain);
-          await popToRoot({
-            clearSearchBar: true,
-          });
-          break;
-
-        default:
-          setState((prevState: State) => {
-            return { ...prevState, isLoading: false };
-          });
-          break;
+      if (response.type==="error") {
+        await showToast(Toast.Style.Success, "Domain Deleted", "DOMAIN: " + domain);
+        getFromApi();
       }
+      setIsLoading(false);
     }
   };
 
   const updateDomainSettings = async (domain: Domain) => {
-    setState((prevState: State) => {
-      return { ...prevState, isLoading: true };
-    });
+    setIsLoading(true);
     if (domain.isShared) {
       await showToast(Toast.Style.Failure, "Purelymail Error", "Can't Edit Shared Domains.");
     } else {
@@ -124,35 +81,39 @@ export default function ListDomains() {
         },
       });
     }
-    setState((prevState) => {
-      return { ...prevState, isLoading: false };
-    });
+    setIsLoading(false);
   };
 
-  const toggleSharedDomains = () => {
-    setState((prevState) => {
-      return { ...prevState, listSharedDomains: !state.listSharedDomains };
-    });
-  };
+  const filteredDomains = !domains ? [] : domains.filter((domain) => {
+    if (filter==="domains_shared") return domain.isShared;
+    if (filter==="domains_custom") return !domain.isShared;
+    return domain;
+  });
 
-  return state.error ? (
-    <ErrorComponent error={state.error} />
+  const domainsTitle = `${filteredDomains.length} of ${domains?.length || 0} domains`;
+
+  return error ? (
+    <ErrorComponent error={error} />
   ) : (
     <List
-      isLoading={state.domains === undefined || state.isLoading}
+      isLoading={isLoading}
       searchBarPlaceholder="Search for domain..."
-      isShowingDetail={state.showDetails}
+      isShowingDetail={isShowingDetails}
+      searchBarAccessory={<List.Dropdown tooltip="Select Domain Type" onChange={(newValue) => setFilter(newValue)}>
+        <List.Dropdown.Item title="All Domains" value="" icon={Icon.Dot} />
+        <List.Dropdown.Item title="Shared Domains" value="domains_shared" icon={{ source: Icon.Globe, tintColor: Color.Green }} />
+        <List.Dropdown.Item title="Custom Domains" value="domains_custom" icon={{ source: Icon.Globe, tintColor: Color.Red }} />
+      </List.Dropdown>}
     >
-      <List.Section title={`${state.domains?.length || 0} domains`}>
-        {(state.domains || [])
-          .filter((domain) => (state.listSharedDomains ? domain : !domain.isShared))
+      <List.Section title={domainsTitle}>
+        {filteredDomains
           .map((domain) => (
             <List.Item
               key={domain.name}
               icon={getFavicon(`https://${domain.name}`, { fallback: Icon.List })}
               title={domain.name}
               accessories={
-                state.showDetails
+                isShowingDetails
                   ? undefined
                   : [
                       { tag: { value: "isShared", color: domain.isShared ? Color.Green : Color.Red } },
@@ -189,6 +150,7 @@ export default function ListDomains() {
                     onAction={() => handleDelete(domain.name)}
                     icon={Icon.DeleteDocument}
                     shortcut={{ modifiers: ["cmd"], key: "d" }}
+                    style={Action.Style.Destructive}
                   />
                 </ActionPanel>
               }
@@ -229,6 +191,8 @@ export default function ListDomains() {
               }
             />
           ))}
+      </List.Section>
+      <List.Section title="Commands">
         <List.Item
           title="Add New Domain"
           icon={{ source: Icon.Plus }}
@@ -241,15 +205,6 @@ export default function ListDomains() {
                   await launchCommand({ name: "add-domain", type: LaunchType.UserInitiated });
                 }}
               />
-            </ActionPanel>
-          }
-        />
-        <List.Item
-          title="Toggle Shared Domains"
-          icon={Icon.Repeat}
-          actions={
-            <ActionPanel>
-              <Action title="Toggle Shared Domains" onAction={toggleSharedDomains} />
             </ActionPanel>
           }
         />